+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - feat: air quality sensor matter device
 - accessory tile + modal improvements/fixes
 - improve keyboard loading on terminals
+- apply save button to accessory info modal
 
 ### Other Changes